+++ air/dockerized_air.sh
@@ -7,6 +7,7 @@ cd $(dirname $0)
 . ./config/config.sh
 
 function stop_docker_services {
+  site/container.sh stop&
   frontend/container.sh stop&
   backend/container.sh stop&
   router/container.sh stop&
@@ -26,13 +27,14 @@ case "$1" in
 
   start)
     ./start_dependencies.sh --no-router
+    start_docker_service site
     start_docker_service frontend
     start_docker_service backend
     start_docker_service router
     start_docker_service balancer
 
     printf "\nYou can access the site at:\n"
-    printf "  https://frontend.air-local:$(get_tcp_port prod balancer/https)\n\n"
+    printf "  https://insights.air-local:$(get_tcp_port prod balancer/https)\n\n"
     ;;
 
   *)