+++ air/site/lib/air.ex
@@ -5,21 +5,9 @@ defmodule Air do
   # See http://elixir-lang.org/docs/stable/elixir/Application.html
   # for more information on OTP Applications
   def start(_type, _args) do
-    import Supervisor.Spec, warn: false
-
-    children = [
-      # Start the endpoint when the application starts
-      supervisor(Air.Endpoint, []),
-      # Start the Ecto repository
-      supervisor(Air.Repo, []),
-      # Here you could define other workers and supervisors as children
-      # worker(Air.Worker, [arg1, arg2, arg3]),
-    ]
-
-    # See http://elixir-lang.org/docs/stable/elixir/Supervisor.html
-    # for other strategies and supported options
-    opts = [strategy: :one_for_one, name: Air.Supervisor]
-    Supervisor.start_link(children, opts)
+    Air.Repo.configure()
+    Air.Endpoint.configure()
+    Air.Supervisor.start_link()
   end
 
   # Tell Phoenix to update the endpoint configuration