+++ air/site/lib/air/supervisor.ex
@@ -1,0 +1,36 @@
+defmodule Air.Supervisor do
+  @moduledoc false
+
+  def start_link,
+    do: Supervisor.start_link(children, strategy: :one_for_one, name: Air.Supervisor)
+
+  # Conditional definition of top-level processes, since we don't want to run
+  # all of them in the test environment.
+  case Mix.env do
+    :test -> defp children, do: common_processes()
+    :dev -> defp children, do: common_processes() ++ system_processes()
+    :prod -> defp children, do: common_processes() ++ system_processes()
+  end
+
+  # Processes which need to run in all environments (including :test)
+  defp common_processes do
+    import Supervisor.Spec, warn: false
+
+    [
+      supervisor(Air.Repo, []),
+      worker(Air.Repo.Migrator, [], restart: :transient),
+      Air.Endpoint.supervisor_spec()
+    ]
+  end
+
+  unless Mix.env == :test do
+    # Processes which we don't want to start in the test environment
+    defp system_processes do
+      import Supervisor.Spec, warn: false
+
+      [
+        Air.PeerDiscovery.supervisor_spec()
+      ]
+    end
+  end
+end