+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 This document serves to describe breaking changes and provide upgrade hints when major changes are introduced. When you're creating a pull with some major changes, please add brief upgrade instructions here.
 
+## Dockerization of insights
+
+- The role which accesses the test database has been changed, so your local tests will probably break. To fix this, start the docker database container, drop `air_test` database, and finally restart the database container.
+
 ## Running air (insights) through the router
 
 - New pseudo local site has been introduced: `insights.air-local`. You need to add the corresponding entry to your `/etc/hosts`.